requirements.txt
paho-mqtt
black
flask[async]
pymongo
configparser
requests